fixing some bugs
