+++ src/App.js
@@ -6,7 +6,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 
 function App() {
   return (
-    <div className="App">    
+    <div className="App container">    
       <CollapsePanel />      
     </div>
   );
