+++ src/components/CollapsePanel/CollapsePanel.js
@@ -1,28 +1,27 @@
 import React, { useState } from 'react';
 import { Collapse } from 'reactstrap';
 import { data } from '../../assets/data';
-import Page from '../Carousel/Carousel';
+import Slide from '../Slide/Slide';
 import './CollapsePanel.css';
 import { ReactComponent as RotateCaret } from '../../assets/rotated_caret.svg';
 import { ReactComponent as File } from '../../assets/file.svg';
 
 const CollapsePanel = () => {
   const [isOpen, setIsOpen] = useState(true);
-  const toggle = () => setIsOpen(!isOpen); 
-  console.log('isopen',isOpen); 
+  const toggle = () => setIsOpen(!isOpen);   
 
   return (
-    <div className='d-block container box'>      
-          <a className='mynav mybg' onClick={toggle} data-toggle="collapse" data-target="#collapseExampleArea" href="#collapseExampleArea" aria-expanded="false" >
+    <div className='d-block box'>      
+          <a className='mynav mybg' onClick={toggle} >
               <div className=' myfont' >
                   <File />{data[0].title}                
               </div>              
               <div className={`expand_caret ${isOpen?'rotatecaret':null}`}>
-                  <RotateCaret/ >
+                  <RotateCaret />
               </div>
           </a>    
       <Collapse isOpen={isOpen} >
-        <Page content={ data[0].content } />
+        <Slide />
       </Collapse>        
     </div>
   );
